Refactor delete task command
