+++ src/main/java/seedu/oneline/logic/commands/DeleteTaskCommand.java
@@ -1,0 +1,51 @@
+package seedu.oneline.logic.commands;
+
+import seedu.oneline.commons.core.Messages;
+import seedu.oneline.commons.core.UnmodifiableObservableList;
+import seedu.oneline.commons.exceptions.IllegalCmdArgsException;
+import seedu.oneline.commons.exceptions.IllegalValueException;
+import seedu.oneline.logic.parser.Parser;
+import seedu.oneline.model.task.ReadOnlyTask;
+import seedu.oneline.model.task.UniqueTaskList.TaskNotFoundException;
+
+public class DeleteTaskCommand extends DeleteCommand {
+    public final int targetIndex;
+
+    public DeleteTaskCommand(int targetIndex) {
+        this.targetIndex = targetIndex;
+    }
+    
+    public static DeleteTaskCommand createFromArgs(String args) throws IllegalCmdArgsException {
+        Integer index = null;
+        try {
+            index = Parser.getIndexFromArgs(args);
+        } catch (IllegalValueException e) {
+            throw new IllegalCmdArgsException(Messages.getInvalidCommandFormatMessage(MESSAGE_USAGE));
+        }
+        if (index == null) {
+            throw new IllegalCmdArgsException(Messages.getInvalidCommandFormatMessage(MESSAGE_USAGE));
+        }
+        return new DeleteTaskCommand(index);
+    }
+    
+    @Override
+    public CommandResult execute() {
+
+        UnmodifiableObservableList<ReadOnlyTask> lastShownList = model.getFilteredTaskList();
+
+        if (lastShownList.size() < targetIndex) {
+            indicateAttemptToExecuteIncorrectCommand();
+            return new CommandResult(Messages.MESSAGE_INVALID_TASK_DISPLAYED_INDEX);
+        }
+
+        ReadOnlyTask taskToDelete = lastShownList.get(targetIndex - 1);
+
+        try {
+            model.deleteTask(taskToDelete);
+        } catch (TaskNotFoundException pnfe) {
+            assert false : "The target task cannot be missing";
+        }
+
+        return new CommandResult(String.format(MESSAGE_DELETE_TASK_SUCCESS, taskToDelete));
+    }
+}
